Rename LogMessage macro to avoid naming conflict with glog.
The Google Logging Library (glog) is a widely used lib. Just as `jetson-utils`, it defines a macro `LogMessage`. In `jetson-utils` this macro is defined and used only within `logging.h`, for defining the actual macros for developers (such as `LogWarning`), and thus may safely be renamed.
This renaming avoid conflicts with glog.